Kim Kyong-hui
Kim Kyong-hui (Korean: 김경희; born 30 May 1946) is the aunt of current North Korean leader, Kim Jong-un. She is the daughter of the founding North Korean leader Kim Il-sung and the sister of the late leader Kim Jong-il. She currently serves as Secretary for Organization of the Workers' Party of Korea. An important member of Kim Jong-il's inner circle of trusted friends and advisors,[2] she was director of the WPK Light Industry Department from 1988 to 2012.[3] She was married to Jang Sung-taek, who was executed in December 2013 in Pyongyang, after being charged with treason and corruption.[4]