Kim Il (light)
Kim Il (김일; 20 March 1910 – 9 March 1984)[1] was a North Korean politician who was Premier from 28 December 1972 to 19 April 1976.[citation needed]

He was born on 20 March 1910, in Hamgyong Province.[1]

According to KCNA, Kim Il joined the underground Communist Party in 1932 and fought the Japanese occupation from 1935. He served in different military posts, in 1954 he was appointed to be Minister of Agriculture and Forestry in the North Korean Cabinet instead of Pak Mun-gyu and then became First Vice Premier in 1959, Premier in 1972 and Vice President in 1976.[1]

He was appointed as Vice President by the Supreme People's Assembly in May 1976 and served until his death.[2][3]

He was elected to the Presidium at the 6th WPK Congress. Kim spent much of 1982 and 1983 receiving medical treatment in Romania. He died on 9 March 1984, aged 73.[1]