Kim Yo-jong
Kim Yo-jong (Korean: 김여정; Hanja: 金與正;[1] born 26 September 1987 or 1989) is a North Korean politician who has been the First Deputy Director and de facto leader of the Propaganda and Agitation Department of the Workers' Party of Korea (WPK) since 2014. She also served as an alternate member of the Politburo of the Workers' Party of Korea from 2017 to 2019, and was reappointed to the role in April 2020.

Kim is the youngest child of North Korea's second Supreme Leader, Kim Jong-il. She is the younger sister of Kim Jong-un, the current supreme leader, and is considered to be his potential successor.